# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crossword Helper API
  version: 2.0.0
  description: |
    REST API for crossword construction assistance tools including pattern matching,
    grid autofill, theme word placement, and wordlist management.

    **Architecture**: Flask backend delegating to CLI via subprocess for single source of truth.

    **Features**:
    - Pattern search across wordlists
    - Grid numbering and validation
    - Entry normalization (crossword conventions)
    - CSP-based autofill with progress tracking
    - Theme word placement suggestions
    - Strategic black square placement
    - Pause/resume autofill with user edits
    - Wordlist management and import

  contact:
    name: Crossword Helper Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server (this is a local-only tool; no other deployment exists)
...
paths:
...
  /grid/suggest-black-square:
    post:
      tags:
        - Grid Helpers
      summary: Suggest black square placements
      description: |
        Suggest optimal black square placements to split problematic slots.
        All suggestions maintain rotational symmetry and standard word counts.
      operationId: suggestBlackSquare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlackSquareRequest'
      responses:
        '200':
          description: Suggestions generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlackSquareResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /grid/apply-black-squares:
    post:
      tags:
        - Grid Helpers
      summary: Apply black square pair
      description: |
        Apply a symmetric black square pair to the grid.
        Both positions are updated simultaneously to maintain symmetry.
      operationId: applyBlackSquares
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grid
                - primary
                - symmetric
              properties:
                grid:
                  $ref: '#/components/schemas/Grid'
                primary:
                  $ref: '#/components/schemas/Position'
                symmetric:
                  $ref: '#/components/schemas/Position'
      responses:
        '200':
          description: Black squares applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  grid:
                    $ref: '#/components/schemas/Grid'
                  applied:
                    type: boolean
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /grid/validate:
    post:
      tags:
        - Grid Helpers
      summary: Validate grid quality
      description: |
        Validate grid quality metrics including word count, black square percentage,
        and symmetry. Provides suggestions for improvement.
      operationId: validateGrid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grid
              properties:
                grid:
                  $ref: '#/components/schemas/Grid'
                grid_size:
                  type: integer
                  default: 15
      responses:
        '200':
          description: Grid validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GridValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    Cell:
      type: object
      description: Single grid cell
      properties:
        letter:
          type: string
          pattern: '^[A-Z]?$'
          description: Letter (A-Z) or empty
        isBlack:
          type: boolean
          description: Whether cell is a black square
        number:
          type: integer
          nullable: true
          description: Clue number (1, 2, 3, ...) or null
        isError:
          type: boolean
          description: Validation error indicator
        isHighlighted:
          type: boolean
          description: UI highlight state
        isThemeLocked:
          type: boolean
          description: Theme-locked cell (cannot be modified by autofill)
    Grid:
      type: array
      description: 2D array of cells (rows × columns)
      items:
        type: array
        items:
          $ref: '#/components/schemas/Cell'
    Position:
      type: object
      description: Grid position (row, col)
      required:
        - row
        - col
      properties:
        row:
          type: integer
          minimum: 0
        col:
          type: integer
          minimum: 0
    Slot:
      type: object
      description: Word slot in grid
      required:
        - row
        - col
        - direction
        - length
      properties:
        row:
          type: integer
          description: Starting row
        col:
          type: integer
          description: Starting column
        direction:
          type: string
          enum: [across, down]
        length:
          type: integer
          minimum: 3
        pattern:
          type: string
          description: Current pattern (e.g., "A??LE")
        candidate_count:
          type: integer
          description: Number of valid words

    # ============================================================================
    # Request/Response Schemas
    # ============================================================================
...
    BlackSquareRequest:
      type: object
      required:
        - grid
        - problematic_slot
      properties:
        grid:
          $ref: '#/components/schemas/Grid'
        grid_size:
          type: integer
          default: 15
        problematic_slot:
          $ref: '#/components/schemas/Slot'
        max_suggestions:
          type: integer
          default: 3
          minimum: 1
          maximum: 10
    BlackSquareResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              position:
                type: integer
                description: Position within slot (0-indexed)
              row:
                type: integer
              col:
                type: integer
              score:
                type: integer
                minimum: 0
                maximum: 1000
              reasoning:
                type: string
              left_length:
                type: integer
              right_length:
                type: integer
              symmetric_position:
                $ref: '#/components/schemas/Position'
              new_word_count:
                type: integer
              constraint_reduction:
                type: integer
                minimum: 0
                maximum: 5
        slot_info:
          $ref: '#/components/schemas/Slot'
        grid_size:
          type: integer
        validation:
          $ref: '#/components/schemas/ValidationResult'
    GridValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        word_count:
          type: integer
        black_square_count:
          type: integer
        black_square_percentage:
          type: number
          format: float
        word_count_range:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
        warnings:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
...
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    Error:
      description: |
        Two shapes are actually in use (backend/api/errors.py vs. ad-hoc
        jsonify calls) — there is no single canonical error schema.
      oneOf:
        - $ref: '#/components/schemas/ErrorEnvelope'
        - $ref: '#/components/schemas/ErrorFlat'
    ErrorEnvelope:
      type: object
      description: |
        Shape returned by the shared handle_error() helper — most endpoints
        (core operations, grid/theme mutation routes, pause/resume).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorFlat:
      type: object
      description: |
        Shape returned by endpoints that build jsonify() directly instead of
        going through handle_error() — constraint analysis, wordlist search,
        and several validation branches in grid/theme routes.
      properties:
        error:
          type: string
          description: Error message
        validation:
          $ref: '#/components/schemas/ValidationResult'
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security: []  # No authentication scheme exists — this is a local-only tool